requirements.txt
absl-py==2.4.0
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
Cython==0.29.37
decorator==5.2.1
etils==1.13.0
exceptiongroup==1.3.0
executing==2.2.1
Farama-Notifications==0.0.4
fasteners==0.15
filelock==3.24.2
fonttools==4.59.2
fsspec==2026.2.0
glfw==1.12.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
grpcio==1.74.0
gym==0.25.1
gym-notebook-wrapper==1.3.3
gym-notices==0.1.0
gymnasium==1.2.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.9
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
monotonic==1.6
moviepy==2.2.1
mpmath==1.3.0
mujoco==3.5.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
proglog==0.1.12
prompt_toolkit==3.0.52
protobuf==6.33.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyVirtualDisplay==3.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
stable_baselines3==2.7.1
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
torch==2.10.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.23.0
onnxscript==0.6.2